set up common mocks for ResizeObserver and IntersectionObserver
