+++ src/components/Tabs/Tabs.spec.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from '@testing-library/react';
-import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest';
+import { describe, expect, test, vi } from 'vitest';
 import { DialTabs } from './Tabs';
 import { TabOrientation } from '@/types/tab';
 import * as useIsTabletScreenHook from '@/hooks/use-is-tablet-screen';
@@ -10,30 +10,6 @@ const tabsMock = [
 ];
 
 describe('Dial UI Kit :: DialTabs', () => {
-  let originalResizeObserver: typeof ResizeObserver;
-
-  beforeAll(() => {
-    originalResizeObserver = globalThis.ResizeObserver;
-
-    class ResizeObserverMock implements ResizeObserver {
-      callback: ResizeObserverCallback;
-      constructor(callback: ResizeObserverCallback) {
-        this.callback = callback;
-      }
-      observe = vi.fn();
-      unobserve = vi.fn();
-      disconnect = vi.fn();
-    }
-
-    globalThis.ResizeObserver =
-      ResizeObserverMock as unknown as typeof ResizeObserver;
-  });
-
-  afterAll(() => {
-    vi.restoreAllMocks();
-    globalThis.ResizeObserver = originalResizeObserver;
-  });
-
   test('renders horizontal tabs and handles click', () => {
     const onClick = vi.fn();
     render(<DialTabs tabs={tabsMock} activeTab="tab1" onClick={onClick} />);
